+++ frontend/src/components/open/navigation_arrows.js
