+++ frontend/src/components/show/playlist_show_container.js
@@ -1,6 +1,5 @@
 import { connect } from "react-redux";
 import PlaylistShow from "./playlist_show";
-import { fetchSongs } from '../../actions/'
 
 const mapStateToProp = state => {
   debugger
